+++ src/sql/inspect.py
@@ -422,9 +422,11 @@ class TableDescription(DatabaseInspection):
             warning_background = "white"
             warning_title = ""
 
-        database = ConnectionManager.current.url
-        db_driver = ConnectionManager.current._get_database_information()["driver"]
-        if "duckdb" in database:
+        current = ConnectionManager.current
+        database = current.dialect
+        db_driver = current._get_database_information()["driver"]
+
+        if database and "duckdb" in database:
             db_message = ""
         else:
             db_message = f"""Following statistics are not available in